+++ super_editor/test/super_editor/document_test_tools.dart
@@ -10,6 +10,7 @@ import 'package:super_editor/super_editor_test.dart';
 import 'package:super_editor_markdown/super_editor_markdown.dart';
 import 'package:text_table/text_table.dart';
 
+import '../test_tools.dart';
 import 'test_documents.dart';
 
 /// Extensions on [WidgetTester] that configure and pump [SuperEditor]
@@ -625,33 +626,5 @@ extension TestMessageInterceptor on WidgetTester {
     });
 
     return handler;
-  }
-}
-
-/// A method to handle plaftorm method calls.
-typedef PlatformMethodHandler = Future<ByteData?>? Function(MethodCall methodCall);
-
-/// Configures handlers to intercept platform method calls.
-///
-/// Use [interceptMethod] to configure a handler for a method.
-class PlatformMessageHandler {
-  final _handlers = <String, PlatformMethodHandler>{};
-
-  /// Configures a [handler] to a [method].
-  PlatformMessageHandler interceptMethod(String method, PlatformMethodHandler handler) {
-    _handlers[method] = handler;
-    return this;
-  }
-
-  /// Decodes platform messages and dispatches to the configured handlers.
-  Future<ByteData?>? handleMessage(ByteData? message) async {
-    final methodCall = const JSONMethodCodec().decodeMethodCall(message);
-    final handler = _handlers[methodCall.method];
-
-    if (handler == null) {
-      return null;
-    }
-
-    return await handler(methodCall);
   }
 }
